requirements.txt
streamlit
ged4p
networkx